+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [[PR151]](https://github.com/lanl/singularity-eos/pull/151) fix module install
 
 ### Added (new features/APIs/variables/...)
+- [[PR164]](https://github.com/lanl/singularity-eos/pull/164) provide facilities for an initial temperature guess for PTE
 
 ### Changed (changing behavior/API/variables/...)